+++ README.md
@@ -15,7 +15,7 @@
 <a href='https://play.google.com/store/apps/details?id=black.bracken.picsorter&pcampaignid=pcampaignidMKT-Other-global-all-co-prtnr-py-PartBadge-Mar2515-1'><img alt='Get it on Google Play' src='https://play.google.com/intl/en_us/badges/static/images/badges/en_badge_web_generic.png' width='20%' height='20%'/></a>
 
 # 環境
-* Android 8+ (SDK 26+)
+* Android 9+ (SDK 28+)
 * Kotlin 1.3.72
 
 # スクリーンショット
