+++ app/src/main/java/black/bracken/picsorter/receiver/BootReceiver.kt
@@ -3,8 +3,8 @@ package black.bracken.picsorter.receiver
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import black.bracken.picsorter.service.repository.ImageObserverRepository
-import black.bracken.picsorter.service.repository.SettingsRepository
+import black.bracken.picsorter.data.repository.ImageObserverRepository
+import black.bracken.picsorter.data.repository.SettingsRepository
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import org.koin.core.KoinComponent
 import org.koin.core.inject
