+++ app/src/main/java/black/bracken/picsorter/service/ImageObserver.kt
@@ -4,10 +4,10 @@ import android.app.Service
 import android.content.Intent
 import android.os.FileObserver
 import android.os.IBinder
+import black.bracken.picsorter.data.repository.SettingsRepository
 import black.bracken.picsorter.ext.notificationManager
 import black.bracken.picsorter.notification.DetectionNotification
 import black.bracken.picsorter.notification.ObservingNotification
-import black.bracken.picsorter.service.repository.SettingsRepository
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.awaitClose
 import kotlinx.coroutines.flow.Flow
