+++ app/src/main/java/black/bracken/picsorter/service/repository/ImageObserverSwitcher.kt
@@ -2,21 +2,12 @@ package black.bracken.picsorter.service.repository
 
 import android.content.Context
 import android.content.Intent
+import black.bracken.picsorter.data.repository.ImageObserverRepository
 import black.bracken.picsorter.service.ImageObserver
 import org.koin.core.KoinComponent
 import org.koin.core.inject
 
-interface ImageObserverRepository {
-
-    fun enableObserver()
-
-    fun disableObserver()
-
-    fun verifyWhetherToRun(): Boolean
-
-}
-
-class ImageObserverDataSource : ImageObserverRepository, KoinComponent {
+class ImageObserverSwitcher : ImageObserverRepository, KoinComponent {
 
     private val context by inject<Context>()
