+++ app/src/main/java/black/bracken/picsorter/service/repository/SimpleManipulatingDatabase.kt
@@ -1,6 +1,7 @@
 package black.bracken.picsorter.service.repository
 
 import black.bracken.picsorter.data.SimpleManipulating
+import black.bracken.picsorter.data.repository.SimpleManipulatingRepository
 import black.bracken.picsorter.db.dao.SimpleManipulatingsDao
 import black.bracken.picsorter.db.entity.toEntity
 import kotlinx.coroutines.Dispatchers
@@ -8,19 +9,7 @@ import kotlinx.coroutines.withContext
 import org.koin.core.KoinComponent
 import org.koin.core.inject
 
-interface SimpleManipulatingRepository {
-
-    suspend fun add(manipulating: SimpleManipulating)
-
-    suspend fun removeByName(name: String)
-
-    suspend fun findByName(name: String): SimpleManipulating?
-
-    suspend fun loadAll(): List<SimpleManipulating>
-
-}
-
-class SimpleManipulatingDataSource : SimpleManipulatingRepository, KoinComponent {
+class SimpleManipulatingDatabase : SimpleManipulatingRepository, KoinComponent {
 
     private val manipulatingDao by inject<SimpleManipulatingsDao>()
