+++ app/src/main/java/black/bracken/picsorter/ui/settings/dirchooser/DirectoriesChooserViewModel.kt
@@ -3,7 +3,7 @@ package black.bracken.picsorter.ui.settings.dirchooser
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import black.bracken.picsorter.service.repository.SettingsRepository
+import black.bracken.picsorter.data.repository.SettingsRepository
 
 class DirectoriesChooserViewModel(
     private val settingsRepository: SettingsRepository
