+++ data/src/main/java/black/bracken/picsorter/data/repository/ImageObserverRepository.kt
@@ -1,0 +1,11 @@
+package black.bracken.picsorter.data.repository
+
+interface ImageObserverRepository {
+
+    fun enableObserver()
+
+    fun disableObserver()
+
+    fun verifyWhetherToRun(): Boolean
+
+}
