+++ data/src/main/java/black/bracken/picsorter/data/repository/SettingsRepository.kt
@@ -1,0 +1,15 @@
+package black.bracken.picsorter.data.repository
+
+interface SettingsRepository {
+
+    var shouldRunOnBoot: Boolean
+
+    val directoryPathList: List<String>
+
+    fun addDirectoryPath(path: String)
+
+    fun removeDirectoryPath(path: String)
+
+    fun containsDirectoryPath(path: String): Boolean
+
+}
